moving things around and deleting files that shouldn't be there
